add link to maps app
