+++ frontend/pubspec.yaml
@@ -28,6 +28,7 @@ dependencies:
   graphql_flutter: ^3.1.0
   location_permissions: ^3.0.0
   url_launcher: ^5.7.10
+  maps_launcher: ^1.2.2+2
   flutter_svg: ^0.18.0
 
   # graphQL type generation
